+++ package.json
@@ -7,9 +7,11 @@
     "dev": "vite",
     "build": "vite build",
     "lint": "eslint .",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "start": "vite"
   },
   "dependencies": {
+    "esm": "^3.2.25",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
   },
